feat: separate mongo connection from server.js to run backend tests smoothly
